expose Name constructor in Unison.Name.Internal
