+++ unison-core/src/Unison/Name.hs
@@ -72,6 +72,7 @@ import qualified Data.Text as Text
 import qualified Data.Text.Lazy as Text.Lazy
 import qualified Data.Text.Lazy.Builder as Text (Builder)
 import qualified Data.Text.Lazy.Builder as Text.Builder
+import Unison.Name.Internal (Name (..), fromTextEither, toText)
 import Unison.NameSegment (NameSegment (NameSegment))
 import qualified Unison.NameSegment as NameSegment
 import Unison.Position (Position (..))
@@ -81,39 +82,6 @@ import qualified Unison.Util.List as List
 import qualified Unison.Util.Relation as R
 import Unison.Var (Var)
 import qualified Unison.Var as Var
-
--- | A name is an absolute-or-relative non-empty list of name segments.
-data Name
-  = -- A few example names:
-    --
-    --   "foo.bar"  --> Name Relative ["bar", "foo"]
-    --   ".foo.bar" --> Name Absolute ["bar", "foo"]
-    --   "|>.<|"    --> Name Relative ["<|", "|>"]
-    --   "."        --> Name Relative ["."]
-    --   ".."       --> Name Absolute ["."]
-    --
-    Name
-      -- whether the name is positioned absolutely (to some arbitrary root namespace), or relatively
-      Position
-      -- the name segments in reverse order
-      (List.NonEmpty NameSegment)
-  deriving stock (Eq, Generic)
-
-instance Alphabetical Name where
-  compareAlphabetical n1 n2 =
-    compareAlphabetical (toText n1) (toText n2)
-
-instance IsString Name where
-  fromString =
-    unsafeFromString
-
-instance Ord Name where
-  compare (Name p0 ss0) (Name p1 ss1) =
-    compare ss0 ss1 <> compare p0 p1
-
-instance Show Name where
-  show =
-    Text.unpack . toText
 
 -- | @compareSuffix x y@ compares the suffix of @y@ (in reverse segment order) that is as long as @x@ to @x@ (in reverse
 -- segment order).
@@ -261,7 +229,7 @@ setPosition pos (Name _ ss) =
 -- | Compute the "parent" of a name, unless the name is only a single segment, in which case it has no parent.
 --
 -- >>> parent "a.b.c"
--- Just "b.c"
+-- Just "a.b"
 --
 -- >>> parent ".a.b.c"
 -- Just ".a.b"
@@ -479,24 +447,6 @@ toString :: Name -> String
 toString =
   Text.unpack . toText
 
--- | Convert a name to a string representation.
-toText :: Name -> Text
-toText (Name pos (x0 :| xs)) =
-  build (buildPos pos <> foldr step mempty xs <> NameSegment.toTextBuilder x0)
-  where
-    step :: NameSegment -> Text.Builder -> Text.Builder
-    step x acc =
-      acc <> NameSegment.toTextBuilder x <> "."
-
-    build :: Text.Builder -> Text
-    build =
-      Text.Lazy.toStrict . Text.Builder.toLazyText
-
-    buildPos :: Position -> Text.Builder
-    buildPos = \case
-      Absolute -> "."
-      Relative -> ""
-
 -- | Convert a name to a string representation, then parse that as a var.
 toVar :: Var v => Name -> v
 toVar =
@@ -559,25 +509,11 @@ commonPrefix :: Name -> Name -> [NameSegment]
 commonPrefix x@(Name p1 _) y@(Name p2 _)
   | p1 /= p2 = []
   | otherwise =
-    commonPrefix' (toList $ segments x) (toList $ segments y)
+      commonPrefix' (toList $ segments x) (toList $ segments y)
   where
     commonPrefix' (a : as) (b : bs)
       | a == b = a : commonPrefix' as bs
     commonPrefix' _ _ = []
-
--- | Unsafely parse a name from a string literal.
---
--- See 'unsafeFromText'.
-unsafeFromString :: String -> Name
-unsafeFromString =
-  unsafeFromText . Text.pack
-
--- | Unsafely parse a name from a string literal.
---
--- Performs very minor validation (a name can't be empty, nor contain a '#' character [at least currently?]) but makes
--- no attempt at rejecting bogus names like "foo...bar...baz".
-unsafeFromText :: HasCallStack => Text -> Name
-unsafeFromText = either (error . Text.unpack) id . fromTextEither
 
 -- | Parse a name from a string literal.
 --
@@ -585,32 +521,6 @@ unsafeFromText = either (error . Text.unpack) id . fromTextEither
 -- no attempt at rejecting bogus names like "foo...bar...baz".
 fromText :: Text -> Maybe Name
 fromText = eitherToMaybe . fromTextEither
-
--- | Unsafely parse a name from a string literal.
---
--- Performs very minor validation (a name can't be empty, nor contain a '#' character [at least currently?]) but makes
--- no attempt at rejecting bogus names like "foo...bar...baz".
-fromTextEither :: Text -> Either Text Name
-fromTextEither = \case
-  "" -> Left "empty name"
-  "." -> Right $ Name Relative ("." :| [])
-  ".." -> Right $ Name Absolute ("." :| [])
-  name
-    | Text.any (== '#') name -> Left ("not a name: " <> tShow name)
-    | Text.head name == '.' -> Name Absolute <$> (go (Text.tail name))
-    | otherwise -> Name Relative <$> go name
-  where
-    go :: Text -> Either Text (List.NonEmpty NameSegment)
-    go name =
-      if ".." `Text.isSuffixOf` name
-        then Right $ "." :| split (Text.dropEnd 2 name)
-        else case split name of
-          [] -> Left "empty name"
-          s : ss -> Right $ s :| ss
-
-    split :: Text -> [NameSegment]
-    split =
-      reverse . map NameSegment . Text.split (== '.')
 
 -- | Unsafely parse a name from a var, by first rendering the var as a string.
 --
